add: integrate firebase bucket

+++ src/app/api/product/delete-product.ts
@@ -1,5 +1,6 @@
-import { db } from "@/lib/firebase";
-import { doc, deleteDoc } from "firebase/firestore";
+import { db, storage } from "@/lib/firebase";
+import { doc, deleteDoc, getDoc } from "firebase/firestore";
+import { ref, deleteObject } from "firebase/storage";
 
 export const deleteProductById = async (
   productId: string,
@@ -7,12 +8,24 @@ export const deleteProductById = async (
   try {
     const productRef = doc(db, "product", productId);
 
+    const productSnap = await getDoc(productRef);
+    if (!productSnap.exists()) {
+      console.error(`Produk dengan ID ${productId} tidak ditemukan.`);
+      return false;
+    }
+
+    const productData = productSnap.data();
+
+    if (productData.image_url) {
+      const imageRef = ref(storage, productData.image_url);
+      await deleteObject(imageRef);
+    }
+
     await deleteDoc(productRef);
 
-    console.log(`Product with ID ${productId} has been deleted successfully.`);
     return true;
   } catch (error) {
-    console.error("Error deleting product:", error);
+    console.error("Gagal menghapus produk:", error);
     return false;
   }
 };
